net/wifibox-alpine: Move away from the use of Alpine Linux mirrors
https://mirror.fel.cvut.cz/alpine/snapshots/ has become unavailable
which was used as a source for getting Alpine Linux packages from
a given point of time.  APK files are meant to be associated with
a rolling package repository and as a consequence, they are moving
targets.  This does not go well with our use case here.

It is not sure if this is going to be restored, that was probably
served on a best-effort basis.

Switch to a more reliable solution and manually mirror the files
to GitHub for each release.  Existing builds should just keep
working, it makes difference only for new builds where optional
APKs are used that were not cached by the FreeBSD distfile mirror.

Fixes freebsd-wifibox#101 [1]

[1] https://github.com/pgj/freebsd-wifibox/issues/101