Enable non-interactive mode for ports.
By default, unconfigured ports launch the dialog module to offer
choices for the user and wait for them to complete the
configuration.  But if ports are run inside a VM and it does not
disable its terminal, the build process may hang indefinitely
on the first run because it will be waiting for the user input.
Mitigate this situation by switching over to non-interactive mode.